Move DictStateReader to runtime/starknet/state.rs
