+++ crates/cheatnet/src/runtime_extensions/forge_runtime_extension/cheatcodes/deploy.rs
@@ -10,7 +10,7 @@ use blockifier::execution::entry_point::{
     ConstructorContext, EntryPointExecutionContext, ExecutionResources,
 };
 use blockifier::execution::execution_utils::felt_to_stark_felt;
-use runtime::context::{build_block_context, build_transaction_context};
+use runtime::starknet::context::{build_block_context, build_transaction_context};
 use runtime::EnhancedHintError;
 use std::sync::Arc;
