+++ crates/runtime/src/lib.rs
@@ -34,7 +34,7 @@ use cairo_vm::vm::errors::memory_errors::MemoryError;
 use starknet_api::StarknetApiError;
 use thiserror::Error;
 
-pub mod context;
+pub mod starknet;
 pub mod utils;
 
 pub trait SyscallPtrAccess {
