+++ crates/runtime/src/starknet/context.rs
@@ -4,13 +4,9 @@ use std::sync::Arc;
 
 use blockifier::block_context::{FeeTokenAddresses, GasPrices};
 
-use blockifier::state::cached_state::ContractStorageKey;
-use blockifier::state::errors::StateError;
-use blockifier::state::state_api::{StateReader, StateResult};
 use blockifier::transaction::objects::{CommonAccountFields, CurrentAccountTransactionContext};
 use blockifier::{
-    abi::constants, block_context::BlockContext, execution::contract_class::ContractClass,
-    transaction::objects::AccountTransactionContext,
+    abi::constants, block_context::BlockContext, transaction::objects::AccountTransactionContext,
 };
 
 use cairo_vm::vm::runners::builtin_runner::{
@@ -22,11 +18,9 @@ use serde::{Deserialize, Serialize};
 use starknet_api::data_availability::DataAvailabilityMode;
 
 use starknet_api::block::{BlockNumber, BlockTimestamp};
-use starknet_api::core::CompiledClassHash;
-use starknet_api::state::StorageKey;
 use starknet_api::transaction::{Resource, ResourceBounds, ResourceBoundsMapping};
 use starknet_api::{
-    core::{ChainId, ClassHash, ContractAddress, Nonce, PatriciaKey},
+    core::{ChainId, ContractAddress, Nonce, PatriciaKey},
     hash::{StarkFelt, StarkHash},
     patricia_key,
     transaction::{TransactionHash, TransactionSignature, TransactionVersion},
@@ -150,67 +144,5 @@ impl Default for BlockInfo {
             timestamp: BlockTimestamp::default(),
             sequencer_address: ContractAddress(patricia_key!(SEQUENCER_ADDRESS)),
         }
-    }
-}
-
-/// A simple implementation of `StateReader` using `HashMap`s as storage.
-#[derive(Debug, Default)]
-pub struct DictStateReader {
-    pub storage_view: HashMap<ContractStorageKey, StarkFelt>,
-    pub address_to_nonce: HashMap<ContractAddress, Nonce>,
-    pub address_to_class_hash: HashMap<ContractAddress, ClassHash>,
-    pub class_hash_to_class: HashMap<ClassHash, ContractClass>,
-    pub class_hash_to_compiled_class_hash: HashMap<ClassHash, CompiledClassHash>,
-}
-
-impl StateReader for DictStateReader {
-    fn get_storage_at(
-        &mut self,
-        contract_address: ContractAddress,
-        key: StorageKey,
-    ) -> StateResult<StarkFelt> {
-        let contract_storage_key = (contract_address, key);
-        self.storage_view
-            .get(&contract_storage_key)
-            .copied()
-            .ok_or(StateError::StateReadError(format!(
-                "Unable to get storage at address: {contract_address:?} and key: {key:?} form DictStateReader"
-            )))
-    }
-
-    fn get_nonce_at(&mut self, contract_address: ContractAddress) -> StateResult<Nonce> {
-        self.address_to_nonce
-            .get(&contract_address)
-            .copied()
-            .ok_or(StateError::StateReadError(format!(
-                "Unable to get nonce at {contract_address:?} from DictStateReader"
-            )))
-    }
-
-    fn get_class_hash_at(&mut self, contract_address: ContractAddress) -> StateResult<ClassHash> {
-        self.address_to_class_hash
-            .get(&contract_address)
-            .copied()
-            .ok_or(StateError::UnavailableContractAddress(contract_address))
-    }
-
-    fn get_compiled_contract_class(
-        &mut self,
-        class_hash: &ClassHash,
-    ) -> StateResult<ContractClass> {
-        let contract_class = self.class_hash_to_class.get(class_hash).cloned();
-        match contract_class {
-            Some(contract_class) => Ok(contract_class),
-            _ => Err(StateError::UndeclaredClassHash(*class_hash)),
-        }
-    }
-
-    fn get_compiled_class_hash(&mut self, class_hash: ClassHash) -> StateResult<CompiledClassHash> {
-        let compiled_class_hash = self
-            .class_hash_to_compiled_class_hash
-            .get(&class_hash)
-            .copied()
-            .unwrap_or_default();
-        Ok(compiled_class_hash)
     }
 }
